requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
idna==3.7
inotify==0.2.10
inotify-simple==1.3.5
kiwisolver==1.4.5
matplotlib==3.9.0
nose==1.3.7
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
pycdata @ file:///home/vc9387/Documents/code/pycdata/pycdata
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyvista==0.43.10
requests==2.32.3
scooby==0.10.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
vtk==9.3.0
watchdog==4.0.1
